Make plugin Java 1.4 compatible

+++ trunk/PrintLayoutPlugin/src/de/intevation/printlayout/LayoutFrame.java
@@ -296,7 +296,7 @@ implements   PickingInteractor.PickingListener
 
 		Integer mnemonic = I18N.getMnemonic(label);
 		if (mnemonic != null)
-			menu.setMnemonic(mnemonic);
+			menu.setMnemonic(mnemonic.intValue());
 
 		return menu;
 	}
